// generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const UserRole = {
  ADMIN: 'ADMIN',
  MANAGER: 'MANAGER',
  EMPLOYEE: 'EMPLOYEE'
} as const

export type UserRole = (typeof UserRole)[keyof typeof UserRole]


export const AccrualScheme = {
  MONTHLY: 'MONTHLY',
  SEMESTER: 'SEMESTER',
  ANNUAL: 'ANNUAL'
} as const

export type AccrualScheme = (typeof AccrualScheme)[keyof typeof AccrualScheme]


export const LeaveRequestStatus = {
  DRAFT: 'DRAFT',
  SUBMITTED: 'SUBMITTED',
  APPROVED: 'APPROVED',
  REJECTED: 'REJECTED',
  CANCELLED: 'CANCELLED'
} as const

export type LeaveRequestStatus = (typeof LeaveRequestStatus)[keyof typeof LeaveRequestStatus]


export const Gender = {
  MALE: 'MALE',
  FEMALE: 'FEMALE',
  OTHER: 'OTHER',
  PREFER_NOT_TO_SAY: 'PREFER_NOT_TO_SAY'
} as const

export type Gender = (typeof Gender)[keyof typeof Gender]


export const AuditActionType = {
  CREATE: 'CREATE',
  UPDATE: 'UPDATE',
  DELETE: 'DELETE',
  APPROVE: 'APPROVE',
  REJECT: 'REJECT',
  CANCEL: 'CANCEL',
  ADJUSTMENT: 'ADJUSTMENT',
  LOGIN: 'LOGIN',
  LOGOUT: 'LOGOUT'
} as const

export type AuditActionType = (typeof AuditActionType)[keyof typeof AuditActionType]


export const RuleType = {
  FIXED_DATE: 'FIXED_DATE',
  NTH_WEEKDAY: 'NTH_WEEKDAY',
  RELATIVE_RULE: 'RELATIVE_RULE',
  MANUAL_ONLY: 'MANUAL_ONLY'
} as const

export type RuleType = (typeof RuleType)[keyof typeof RuleType]


export const OffsetRule = {
  NONE: 'NONE',
  NEXT_MONDAY_IF_WEEKEND: 'NEXT_MONDAY_IF_WEEKEND',
  PREV_FRIDAY_IF_WEEKEND: 'PREV_FRIDAY_IF_WEEKEND'
} as const

export type OffsetRule = (typeof OffsetRule)[keyof typeof OffsetRule]


export const OTPType = {
  FORGOT_PASSWORD: 'FORGOT_PASSWORD',
  TWO_FACTOR: 'TWO_FACTOR',
  LOGIN: 'LOGIN'
} as const

export type OTPType = (typeof OTPType)[keyof typeof OTPType]